Made the themes now responsible for the coloring.
A syntax plugin changing the colors is not a good idea, since there are
many themes out there and none of them would work.

Because of that, I've removed the massive chunk of color-related code
and instead linked the plugin keywords to theme keywords.

I'm not sure if the keywords I used are the correct ones, though, so
that's a thing that might need to be adjusted.